+++ src/models/index.ts
@@ -1944,7 +1944,7 @@ export interface DocumentCreate {
   /** Language code in ISO 639-1 format. Must specify zh-cn or zh-tw for Chinese. */
   language?: string;
   /** If "true", parsing will fail when the uploaded document is duplicate of an existing document, no credits will be consumed. If "false", will parse the document normally whether its a duplicate or not. If not provided, will fallback to the workspace settings. */
-  rejectDuplicates?: string;
+  rejectDuplicates?: boolean;
 }
 
 export interface OrganizationCreate {
@@ -2822,7 +2822,7 @@ export interface AffindaAPICreateDocumentOptionalParams
   /** Language code in ISO 639-1 format. Must specify zh-cn or zh-tw for Chinese. */
   language?: string;
   /** If "true", parsing will fail when the uploaded document is duplicate of an existing document, no credits will be consumed. If "false", will parse the document normally whether its a duplicate or not. If not provided, will fallback to the workspace settings. */
-  rejectDuplicates?: string;
+  rejectDuplicates?: boolean;
 }
 
 /** Contains response data for the createDocument operation. */