+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - rawText is now not nullable
 - Allow custom resume fields to be nullable
+- Allow custom job description fields to be nullable
 
 ### Fixed
 - Allow rejectDuplicates to be null